evaluation/requirements-dev.txt
pytest-mockito==0.0.4
testcontainers==4.9.1
pandas==2.2.3
